Create Prime Numbers in a Range.java
Program to find prime numbers in a given range is discussed here. A number is said to be prime if it is divisible by 1 and the number itself. Program to find prime numbers in a given range using loop

Input & Output Format:

Input consists of two integers. 

First integer corresponds to lower value.

Second integer corresponds to upper value.

Output consists of the prime numbers between the two intervals.

Sample Input: 

1

10

Sample Output:

2

3

5

7