Create Sabrina and Theo’s Game-2
Sabrina and Theo were friends. They felt very bored and decided to play a game with calendar and number of days. Sabrina will give the month and the year as input, and Theo wants to find the number of days in that month. Write a C program using switch case and help Sabrina to find the answer. You should check whether the given year is a leap year. If it is a leap year it will have 29 days in February.

Input & Output Format:

Input consists of two integers.

First input consists of  a month.

Second input consists of a year.

Output consists of a integer.


Sample Input:

3

1996

Sample Output:
Number of days is 31